backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.27.5
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.6.1
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0
